requirements.txt
flet==0.23.2
python-dotenv
pyjwt
bcrypt
psycopg2-binary